feat: implement table reservation logic and endpoints

+++ src/main/java/org/example/api/service/RestaurantTableService.java
@@ -1,21 +1,29 @@
 package org.example.api.service;
 
+import jakarta.persistence.Table;
 import lombok.RequiredArgsConstructor;
 import org.example.api.dto.RestaurantTableResponse;
 import org.example.api.exception.ResourceAlreadyExistsException;
+import org.example.api.exception.ResourceInUseException;
 import org.example.api.exception.ResourceNotFoundException;
+import org.example.api.exception.TableNumberNotUniqueException;
+import org.example.api.model.ReservationStatus;
 import org.example.api.model.RestaurantTable;
+import org.example.api.model.TableType;
+import org.example.api.repository.ReservationRepository;
 import org.example.api.repository.RestaurantTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
 public class RestaurantTableService {
 
     private final RestaurantTableRepository tableRepository;
+    private final ReservationRepository reservationRepository;
 
     public List<RestaurantTableResponse> findAll() {
         return tableRepository.findAll().stream().map(this::mapToResponse).toList();
@@ -23,6 +31,11 @@ public class RestaurantTableService {
 
     @Transactional
     public RestaurantTable create(RestaurantTable table) {
+
+        if (tableRepository.existsRestaurantTableByTableNumber(table.getTableNumber()) && table.getTableType() == TableType.TABLE){
+            throw new TableNumberNotUniqueException("Stolik z numerem " + table.getTableNumber() + " juz istnieje");
+        }
+
         boolean positionTaken = tableRepository.findAll().stream()
                 .anyMatch(t -> t.getRowPosition().equals(table.getRowPosition()) &&
                                t.getColumnPosition().equals(table.getColumnPosition()));
@@ -38,6 +51,14 @@ public class RestaurantTableService {
     public void delete(Long id) {
         if (!tableRepository.existsById(id)) {
             throw new ResourceNotFoundException("Stolik nie istnieje");
+        }
+
+        if (reservationRepository.existsByRestaurantTableIdAndStatusIn(
+                id,
+                Set.of(ReservationStatus.PENDING, ReservationStatus.CONFIRMED, ReservationStatus.IN_PROGRESS))
+        )
+        {
+            throw new ResourceInUseException("Nie można usunąć stolika, który posiada aktywne rezerwacje");
         }
 
         tableRepository.deleteById(id);
